Implementing Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -55,7 +55,9 @@ class QuestionsController extends Controller
      */
     public function show(Question $question)
     {
-        //
+        $question->increment('views');
+
+        return view('questions.show', compact('question'));
     }
 
     /**
@@ -66,6 +68,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question', $question))
+        {
+            abort(403, "Access denied");
+        }
         return view('questions.edit', compact('question'));
 
     }
@@ -79,6 +85,12 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+
+        if(\Gate::denies('update-question', $question))
+        {
+            abort(403, "Access denied");
+        }
+
         $question->update($request->only('title', 'body'));
 
         return redirect()->route('questions.index')->with('success', "Your question has been Edited!");
@@ -93,6 +105,12 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+
+        if(\Gate::denies('delete-question', $question))
+        {
+            abort(403, "Access denied");
+        }
+
         $question->delete();
 
         return redirect()->route('questions.index')->with('success', "Your question has been Deleted!");
